requirements.txt
python-dotenv==1.0.0
openai==1.51.0
pymilvus[lite]>=2.3.4
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
numpy>=1.26.2
requests>=2.31.0